+++ clients/wallet/src/main/java/com/binance/connector/client/wallet/rest/api/AccountApi.java
@@ -50,7 +50,7 @@ public class AccountApi {
 
     private static final String USER_AGENT =
             String.format(
-                    "binance-wallet/6.0.0 (Java/%s; %s; %s)",
+                    "binance-wallet/6.1.0 (Java/%s; %s; %s)",
                     SystemUtil.getJavaVersion(), SystemUtil.getOs(), SystemUtil.getArch());
     private static final boolean HAS_TIME_UNIT = false;