+++ clients/wallet/src/main/java/com/binance/connector/client/wallet/rest/model/GetAssetsThatCanBeConvertedIntoBnbRequest.java
@@ -41,7 +41,7 @@ public class GetAssetsThatCanBeConvertedIntoBnbRequest {
 
     @SerializedName(SERIALIZED_NAME_ACCOUNT_TYPE)
     @jakarta.annotation.Nullable
-    private AccountType accountType = AccountType.SPOT;
+    private AccountType accountType;
 
     public static final String SERIALIZED_NAME_RECV_WINDOW = "recvWindow";