+++ examples/wallet/src/main/java/com/binance/connector/client/wallet/rest/asset/GetSpotAssetTagsExample.java
@@ -6,12 +6,11 @@ import com.binance.connector.client.common.configuration.ClientConfiguration;
 import com.binance.connector.client.common.configuration.SignatureConfiguration;
 import com.binance.connector.client.wallet.rest.WalletRestApiUtil;
 import com.binance.connector.client.wallet.rest.api.WalletRestApi;
-import com.binance.connector.client.wallet.rest.model.OrderType;
-import com.binance.connector.client.wallet.rest.model.QueryUserDelegationHistoryResponse;
+import com.binance.connector.client.wallet.rest.model.GetSpotAssetTagsResponse;
 import java.io.IOException;
 
 /** API examples for AssetApi */
-public class QueryUserDelegationHistoryExample {
+public class GetSpotAssetTagsExample {
     private WalletRestApi api;
 
     public WalletRestApi getApi() {
@@ -27,24 +26,16 @@ public class QueryUserDelegationHistoryExample {
     }
 
     /**
-     * Query User Delegation History(For Master Account) (USER_DATA)
+     * Get Spot Asset Tags (MARKET_DATA)
      *
-     * <p>Query User Delegation History Weight(IP): 60 Security Type: USER_DATA
+     * <p>Get the tags configured for spot-tradable assets. Weight(IP): 100 Security Type:
+     * MARKET_DATA
      *
      * @throws ApiException if the Api call fails
      */
-    public void queryUserDelegationHistoryExample() throws ApiException, IOException {
-        String email = "abc@test.com";
-        Long startTime = 1623319461670L;
-        Long endTime = 1641782889000L;
-        OrderType type = OrderType.DELEGATE;
-        String asset = "BTC";
-        Long current = 1L;
-        Long size = 10L;
-        Long recvWindow = 5000L;
-        ApiResponse<QueryUserDelegationHistoryResponse> response =
-                getApi().queryUserDelegationHistory(
-                                email, startTime, endTime, type, asset, current, size, recvWindow);
+    public void getSpotAssetTagsExample() throws ApiException, IOException {
+        String tag = "Layer1_Layer2,BSC";
+        ApiResponse<GetSpotAssetTagsResponse> response = getApi().getSpotAssetTags(tag);
         System.out.println(response.getData());
     }
 }